refactor: calculate 역할 위임
