+++ src/main/java/chess/domain/board/ChessBoard.java
@@ -7,20 +7,13 @@ import chess.domain.piece.Piece;
 import chess.domain.piece.Team;
 import chess.exception.PieceCannotMoveException;
 
-import java.math.BigDecimal;
-import java.util.Arrays;
-import java.util.List;
 import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 import static chess.domain.piece.PieceType.EMPTY;
 import static chess.domain.piece.PieceType.PAWN;
 
 public class ChessBoard {
-
-    private static final BigDecimal PAWN_REDUCE_NUMBER = new BigDecimal("0.5");
-
+    
     private final PiecePosition piecePosition;
 
     private ChessBoard(final Map<Position, Piece> piecePosition) {
@@ -99,54 +92,12 @@ public class ChessBoard {
         piecePosition.replace(from, to);
     }
 
-    public BigDecimal calculateScore(final Team team) {
-        BigDecimal ordinalScore = getOrdinalScore(team);
-        BigDecimal pawnDifferenceScore = getPawnSubScore(team);
-        return ordinalScore.subtract(pawnDifferenceScore);
-    }
-
-    private BigDecimal getPawnSubScore(final Team team) {
-        List<Position> pawnPositions = toPawnPositions(team);
-        return calculatePawnSubScore(pawnPositions);
-    }
-
-    private BigDecimal getOrdinalScore(final Team team) {
-        return piecePosition.get().values().stream()
-                .filter(piece -> piece.getTeam() == team)
-                .map(piece -> piece.getType().getScore())
-                .reduce(BigDecimal::add)
-                .orElseThrow(() -> new IllegalStateException(team + "은 점수를 계산할 수 없습니다."));
-    }
-
-    private BigDecimal calculatePawnSubScore(final List<Position> pawnPositions) {
-        BigDecimal result = BigDecimal.ZERO;
-
-        result = pawnPositions.stream()
-                .map(Position::getFile)
-                .collect(Collectors.groupingBy(Function.identity(), Collectors.counting()))
-                .values().stream()
-                .filter(count -> count > 1)
-                .map(count -> PAWN_REDUCE_NUMBER.multiply(BigDecimal.valueOf(count)))
-                .reduce(result, BigDecimal::add);
-
-        return result;
-    }
-
-    private List<Position> toPawnPositions(Team team) {
-        return Arrays.stream(File.values())
-                .flatMap(file -> Arrays.stream(Rank.values())
-                        .map(rank -> Position.of(file, rank))
-                ).filter(position -> piecePosition.findPieceByPosition(position).getType() == PAWN)
-                .filter(position -> piecePosition.findPieceByPosition(position).getTeam() == team)
-                .collect(Collectors.toList());
-    }
-
     public boolean isEnd() {
         return piecePosition.isKingAvailable();
     }
 
-    public Map<Position, Piece> getPiecePosition() {
-        return Map.copyOf(piecePosition.get());
+    public PiecePosition getPiecePosition() {
+        return piecePosition;
     }
 
     public Piece get(final Position from) {
