+++ src/main/java/chess/domain/board/PiecePosition.java
@@ -5,18 +5,24 @@ import chess.domain.board.position.Position;
 import chess.domain.board.position.Rank;
 import chess.domain.piece.EmptyPiece;
 import chess.domain.piece.Piece;
+import chess.domain.piece.Team;
 
+import java.math.BigDecimal;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import static chess.domain.piece.PieceType.KING;
+import static chess.domain.piece.PieceType.PAWN;
 
 public class PiecePosition {
 
     private static final int END_OF_SIZE = 1;
     private static final int DEFAULT_KING_COUNT = 2;
+    private static final BigDecimal PAWN_REDUCE_NUMBER = new BigDecimal("0.5");
 
     private final Map<Position, Piece> piecePosition;
 
@@ -87,6 +93,48 @@ public class PiecePosition {
         return files.stream()
                 .flatMap(file -> ranks.stream()
                         .map(rank -> Position.of(file, rank)))
+                .collect(Collectors.toList());
+    }
+
+    public BigDecimal calculateScore(final Team team) {
+        BigDecimal ordinalScore = getOrdinalScore(team);
+        BigDecimal pawnDifferenceScore = getPawnSubScore(team);
+        return ordinalScore.subtract(pawnDifferenceScore);
+    }
+
+    private BigDecimal getPawnSubScore(final Team team) {
+        List<Position> pawnPositions = toPawnPositions(team);
+        return calculatePawnSubScore(pawnPositions);
+    }
+
+    private BigDecimal getOrdinalScore(final Team team) {
+        return piecePosition.values().stream()
+                .filter(piece -> piece.getTeam() == team)
+                .map(piece -> piece.getType().getScore())
+                .reduce(BigDecimal::add)
+                .orElseThrow(() -> new IllegalStateException(team + "은 점수를 계산할 수 없습니다."));
+    }
+
+    private BigDecimal calculatePawnSubScore(final List<Position> pawnPositions) {
+        BigDecimal result = BigDecimal.ZERO;
+
+        result = pawnPositions.stream()
+                .map(Position::getFile)
+                .collect(Collectors.groupingBy(Function.identity(), Collectors.counting()))
+                .values().stream()
+                .filter(count -> count > 1)
+                .map(count -> PAWN_REDUCE_NUMBER.multiply(BigDecimal.valueOf(count)))
+                .reduce(result, BigDecimal::add);
+
+        return result;
+    }
+
+    private List<Position> toPawnPositions(Team team) {
+        return Arrays.stream(File.values())
+                .flatMap(file -> Arrays.stream(Rank.values())
+                        .map(rank -> Position.of(file, rank))
+                ).filter(position -> piecePosition.get(position).getType() == PAWN)
+                .filter(position -> piecePosition.get(position).getTeam() == team)
                 .collect(Collectors.toList());
     }
 }
