+++ src/main/java/chess/dto/ChessBoardDto.java
@@ -17,7 +17,7 @@ public class ChessBoardDto {
     private final List<PieceDto> pieceDtos;
 
     public static ChessBoardDto from(final ChessBoard board) {
-        final List<PieceDto> pieces = pieceToView(board.getPiecePosition());
+        final List<PieceDto> pieces = pieceToView(board.getPiecePosition().get());
         return new ChessBoardDto(pieces);
     }
